requirements.txt
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
eventlet==0.39.1
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
greenlet==3.1.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.2.2
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.2.2
simple-websocket==1.1.0
six==1.17.0
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
